Explore Screen Completed
- Explore screen / Search screen completed
- Custom Collection Layout completed
- Created custom search bar container view

User is able to navigate to the second tab, which is the explore / search screen.
Here, the user is able to explore / search many different posts. Some posts are
normal size while other 'featured' posts are enlarged thanks to the custom
Collection Layout created and applied to the collection view in the Search.storyboard.
A custom search view controller is created and added to the navigation controller.